Clear up
structure project
