+++ NFC.uvproj
@@ -667,6 +667,26 @@
               <FileType>1</FileType>
               <FilePath>.\user\cmd.c</FilePath>
             </File>
+            <File>
+              <FileName>delay.c</FileName>
+              <FileType>1</FileType>
+              <FilePath>.\user\delay.c</FilePath>
+            </File>
+            <File>
+              <FileName>delay.h</FileName>
+              <FileType>5</FileType>
+              <FilePath>.\user\delay.h</FilePath>
+            </File>
+            <File>
+              <FileName>nfc.c</FileName>
+              <FileType>1</FileType>
+              <FilePath>.\nfc.c</FilePath>
+            </File>
+            <File>
+              <FileName>nfc.h</FileName>
+              <FileType>5</FileType>
+              <FilePath>.\user\nfc.h</FilePath>
+            </File>
           </Files>
         </Group>
       </Groups>
