+++ NFC_NFC.dep
@@ -640,7 +640,7 @@ I (.\inc\stm32f0xx_tim.h)(0x4F6C6027)
 I (.\inc\stm32f0xx_usart.h)(0x4F6C6027)
 I (.\inc\stm32f0xx_wwdg.h)(0x4F6C6027)
 I (.\inc\stm32f0xx_misc.h)(0x4F6C6027)
-F (.\user\main.c)(0x56ED2D03)(-c --cpu Cortex-M0 -g -O0 --apcs=interwork -I.\inc -I.\Headers-I C:\Keil_v5\ARM\RV31\INC-I C:\Keil_v5\ARM\PACK\ARM\CMSIS\4.1.1\CMSIS\Include-I C:\Keil_v5\ARM\Inc\ST\STM32F0xx-DUSE_STDPERIPH_DRIVER -DSTM32F0XX -DSTM32F030-o .\debug\main.o --omf_browse .\debug\main.crf --depend .\debug\main.d)
+F (.\user\main.c)(0x56ED591B)(-c --cpu Cortex-M0 -g -O0 --apcs=interwork -I.\inc -I.\Headers-I C:\Keil_v5\ARM\RV31\INC-I C:\Keil_v5\ARM\PACK\ARM\CMSIS\4.1.1\CMSIS\Include-I C:\Keil_v5\ARM\Inc\ST\STM32F0xx-DUSE_STDPERIPH_DRIVER -DSTM32F0XX -DSTM32F030-o .\debug\main.o --omf_browse .\debug\main.crf --depend .\debug\main.d)
 I (.\inc\stm32f0xx.h)(0x4F6C6026)
 I (C:\Keil_v5\ARM\PACK\ARM\CMSIS\4.1.1\CMSIS\Include\core_cm0.h)(0x53688AAC)
 I (C:\Keil_v5\ARM\ARMCC\include\stdint.h)(0x53463A44)
@@ -671,6 +671,7 @@ I (.\inc\stm32f0xx_misc.h)(0x4F6C6027)
 I (C:\Keil_v5\ARM\ARMCC\include\stdio.h)(0x53463A44)
 I (user\cmd.h)(0x56ED0697)
 I (user\delay.h)(0x56ECF47E)
+I (user\nfc.h)(0x56ED5BB0)
 F (.\user\cmd.h)(0x56ED0697)()
 F (.\user\cmd.c)(0x56ECFB5D)(-c --cpu Cortex-M0 -g -O0 --apcs=interwork -I.\inc -I.\Headers-I C:\Keil_v5\ARM\RV31\INC-I C:\Keil_v5\ARM\PACK\ARM\CMSIS\4.1.1\CMSIS\Include-I C:\Keil_v5\ARM\Inc\ST\STM32F0xx-DUSE_STDPERIPH_DRIVER -DSTM32F0XX -DSTM32F030-o .\debug\cmd.o --omf_browse .\debug\cmd.crf --depend .\debug\cmd.d)
 I (user\cmd.h)(0x56ED0697)
@@ -680,5 +681,36 @@ F (.\user\delay.c)(0x56ECF4A6)(-c --cpu Cortex-M0 -g -O0 --apcs=interwork -I.\in
 I (user\delay.h)(0x56ECF47E)
 I (C:\Keil_v5\ARM\ARMCC\include\stdint.h)(0x53463A44)
 F (.\user\delay.h)(0x56ECF47E)()
-F (.\nfc.c)(0x00000000)(-c --cpu Cortex-M0 -g -O0 --apcs=interwork -I.\inc -I.\Headers-I C:\Keil_v5\ARM\RV31\INC-I C:\Keil_v5\ARM\PACK\ARM\CMSIS\4.1.1\CMSIS\Include-I C:\Keil_v5\ARM\Inc\ST\STM32F0xx-DUSE_STDPERIPH_DRIVER -DSTM32F0XX -DSTM32F030-o .\debug\nfc.o --omf_browse .\debug\nfc.crf --depend .\debug\nfc.d)
-F (.\user\nfc.h)(0x56ED2D2C)()
+F (.\user\nfc.h)(0x56ED5BB0)()
+F (.\user\nfc.c)(0x56ED5C46)(-c --cpu Cortex-M0 -g -O0 --apcs=interwork -I.\inc -I.\Headers-I C:\Keil_v5\ARM\RV31\INC-I C:\Keil_v5\ARM\PACK\ARM\CMSIS\4.1.1\CMSIS\Include-I C:\Keil_v5\ARM\Inc\ST\STM32F0xx-DUSE_STDPERIPH_DRIVER -DSTM32F0XX -DSTM32F030-o .\debug\nfc.o --omf_browse .\debug\nfc.crf --depend .\debug\nfc.d)
+I (.\inc\stm32f0xx.h)(0x4F6C6026)
+I (C:\Keil_v5\ARM\PACK\ARM\CMSIS\4.1.1\CMSIS\Include\core_cm0.h)(0x53688AAC)
+I (C:\Keil_v5\ARM\ARMCC\include\stdint.h)(0x53463A44)
+I (C:\Keil_v5\ARM\PACK\ARM\CMSIS\4.1.1\CMSIS\Include\core_cmInstr.h)(0x5302A2D0)
+I (C:\Keil_v5\ARM\PACK\ARM\CMSIS\4.1.1\CMSIS\Include\core_cmFunc.h)(0x5302A2D0)
+I (.\inc\system_stm32f0xx.h)(0x4F6C6026)
+I (.\Headers\stm32f0xx_conf.h)(0x4F6C6028)
+I (.\inc\stm32f0xx_cec.h)(0x4F6C6026)
+I (.\inc\stm32f0xx_crc.h)(0x4F6C6026)
+I (.\inc\stm32f0xx_comp.h)(0x4F6C6026)
+I (.\inc\stm32f0xx_dac.h)(0x4F6C6026)
+I (.\inc\stm32f0xx_dbgmcu.h)(0x4F6C6026)
+I (.\inc\stm32f0xx_dma.h)(0x4F6C6026)
+I (.\inc\stm32f0xx_exti.h)(0x4F6C6026)
+I (.\inc\stm32f0xx_flash.h)(0x4F6C6026)
+I (.\inc\stm32f0xx_gpio.h)(0x4F6C6026)
+I (.\inc\stm32f0xx_syscfg.h)(0x4F6C6027)
+I (.\inc\stm32f0xx_i2c.h)(0x4F6C6026)
+I (.\inc\stm32f0xx_iwdg.h)(0x4F6C6027)
+I (.\inc\stm32f0xx_pwr.h)(0x4F6C6027)
+I (.\inc\stm32f0xx_rcc.h)(0x4F6C6027)
+I (.\inc\stm32f0xx_rtc.h)(0x4F6C6027)
+I (.\inc\stm32f0xx_spi.h)(0x4F6C6027)
+I (.\inc\stm32f0xx_tim.h)(0x4F6C6027)
+I (.\inc\stm32f0xx_usart.h)(0x4F6C6027)
+I (.\inc\stm32f0xx_wwdg.h)(0x4F6C6027)
+I (.\inc\stm32f0xx_misc.h)(0x4F6C6027)
+I (C:\Keil_v5\ARM\ARMCC\include\stdio.h)(0x53463A44)
+I (user\nfc.h)(0x56ED5BB0)
+I (user\cmd.h)(0x56ED0697)
+I (user\delay.h)(0x56ECF47E)
