+++ NFC_uvproj.bak
@@ -657,6 +657,16 @@
               <FileType>1</FileType>
               <FilePath>.\user\main.c</FilePath>
             </File>
+            <File>
+              <FileName>cmd.h</FileName>
+              <FileType>5</FileType>
+              <FilePath>.\user\cmd.h</FilePath>
+            </File>
+            <File>
+              <FileName>cmd.c</FileName>
+              <FileType>1</FileType>
+              <FilePath>.\user\cmd.c</FilePath>
+            </File>
           </Files>
         </Group>
       </Groups>
