+++ debug/NFC.build_log.htm
@@ -4,35 +4,40 @@
 <h1>µVision Build Log</h1>
 <h2>Project:</h2>
 D:\WorkSpace\NFC_Project\NFC\NFC.uvproj
-Project File Date:  03/18/2016
+Project File Date:  03/19/2016
 
 <h2>Output:</h2>
-Build target 'NFC'
-compiling main.c...
-user\main.c(26): warning:  #177-D: variable "SBUFF" was declared but never referenced
-        uint16_t SBUFF[2];
-user\main.c: 1 warning, 0 errors
-compiling nfc.c...
-nfc.c: Error:  #5: cannot open source input file "nfc.c": No such file or directory
-nfc.c: 0 warnings, 1 error
-".\debug\NFC.axf" - 1 Error(s), 1 Warning(s).
-Target not created
-compiling main.c...
-user\main.c(26): warning:  #177-D: variable "SBUFF" was declared but never referenced
-        uint16_t SBUFF[2];
-user\main.c: 1 warning, 0 errors
-"user\main.c" - 0 Error(s), 1 Warning(s).
-.c...
+Rebuild target 'NFC'
+assembling startup_stm32f030.s...
+compiling stm32f0xx_adc.c...
+compiling stm32f0xx_cec.c...
+compiling stm32f0xx_comp.c...
+compiling stm32f0xx_crc.c...
+compiling stm32f0xx_dac.c...
+compiling stm32f0xx_dbgmcu.c...
+compiling stm32f0xx_dma.c...
+compiling stm32f0xx_exti.c...
+compiling stm32f0xx_flash.c...
+compiling stm32f0xx_gpio.c...
+compiling stm32f0xx_i2c.c...
+compiling stm32f0xx_iwdg.c...
+compiling stm32f0xx_misc.c...
+compiling stm32f0xx_pwr.c...
+compiling stm32f0xx_rcc.c...
+compiling stm32f0xx_rtc.c...
+compiling stm32f0xx_spi.c...
+compiling stm32f0xx_syscfg.c...
 compiling stm32f0xx_tim.c...
 compiling stm32f0xx_usart.c...
 compiling stm32f0xx_wwdg.c...
 compiling system_stm32f0xx.c...
 compiling main.c...
-user\main.c(26): warning:  #177-D: variable "SBUFF" was declared but never referenced
+user\main.c(20): warning:  #177-D: variable "SBUFF" was declared but never referenced
         uint16_t SBUFF[2];
 user\main.c: 1 warning, 0 errors
 compiling cmd.c...
 compiling delay.c...
+compiling nfc.c...
 linking...
 Program Size: Code=5380 RO-data=212 RW-data=44 ZI-data=1708  
 FromELF: creating hex file...
@@ -70,11 +75,37 @@ Erase Done.
 Programming Done.
 Verify OK.
 Application running ...
-g ...
-chpoints:          2
-JTAG speed: 1000 kHz
- 
-Erase Done.
-Programming Done.
-Verify OK.
-Application running ...
+iph_USART1, ENABLE);
+user\nfc.c(29): error:  #20: identifier "RCC_APB2Periph_USART1" is undefined
+      RCC_APB2PeriphClockCmd(RCC_APB2Periph_USART1, ENABLE);
+user\nfc.c(31): error:  #20: identifier "GPIO_Pin_2" is undefined
+      GPIO_InitStructure.GPIO_Pin = GPIO_Pin_2 | GPIO_Pin_3;
+user\nfc.c(31): error:  #20: identifier "GPIO_Pin_3" is undefined
+      GPIO_InitStructure.GPIO_Pin = GPIO_Pin_2 | GPIO_Pin_3;
+user\nfc.c(32): error:  #20: identifier "GPIO_Mode_AF" is undefined
+      GPIO_InitStructure.GPIO_Mode = GPIO_Mode_AF;
+user\nfc.c(33): error:  #20: identifier "GPIO_OType_PP" is undefined
+      GPIO_InitStructure.GPIO_OType = GPIO_OType_PP;
+user\nfc.c(34): error:  #20: identifier "GPIO_PuPd_UP" is undefined
+      GPIO_InitStructure.GPIO_PuPd = GPIO_PuPd_UP;
+user\nfc.c(35): error:  #20: identifier "GPIO_Speed_50MHz" is undefined
+      GPIO_InitStructure.GPIO_Speed = GPIO_Speed_50MHz;
+user\nfc.c(36): warning:  #223-D: function "GPIO_Init" declared implicitly
+      GPIO_Init(GPIOA, &GPIO_InitStructure);
+user\nfc.c(36): error:  #20: identifier "GPIOA" is undefined
+      GPIO_Init(GPIOA, &GPIO_InitStructure);
+user\nfc.c(38): warning:  #223-D: function "GPIO_PinAFConfig" declared implicitly
+      GPIO_PinAFConfig(GPIOA, GPIO_PinSource2, GPIO_AF_1);
+user\nfc.c(38): error:  #20: identifier "GPIO_PinSource2" is undefined
+      GPIO_PinAFConfig(GPIOA, GPIO_PinSource2, GPIO_AF_1);
+user\nfc.c(38): error:  #20: identifier "GPIO_AF_1" is undefined
+      GPIO_PinAFConfig(GPIOA, GPIO_PinSource2, GPIO_AF_1);
+user\nfc.c(39): error:  #20: identifier "GPIO_PinSource3" is undefined
+      GPIO_PinAFConfig(GPIOA, GPIO_PinSource3, GPIO_AF_1);
+user\nfc.c(43): error:  #20: identifier "USART_HardwareFlowControl_None" is undefined
+      USART_InitStructure.USART_HardwareFlowControl = USART_HardwareFlowControl_None;
+user\nfc.c(44): error:  #20: identifier "USART_Mode_Rx" is undefined
+      USART_InitStructure.USART_Mode = USART_Mode_Rx | USART_Mode_Tx;
+user\nfc.c: 4 warnings, 30 errors
+".\debug\NFC.axf" - 30 Error(s), 4 Warning(s).
+Target not created
