+++ debug/NFC.hex
@@ -19,7 +19,7 @@
 :10011000704700000023002400250026103A01D378
 :1001200078C1FBD8520700D330C100D50B607047AF
 :100130001FB5C046C0461FBD10B510BD01F0FCF98B
-:100140001146FFF7F5FF01F070F801F014FA03B45F
+:100140001146FFF7F5FF00F0CFFF01F014FA03B4FA
 :10015000FFF7F2FF03BC01F019FA000007488047DF
 :1001600007480047FEE7FEE7FEE7FEE7FEE7FEE79B
 :1001700004480549054A064B70470000BF0F0008B8
@@ -269,54 +269,54 @@
 :1010B000F021084000090B490D5C08480068E84031
 :1010C00006490860F8BD0000001002400020024000
 :1010D0000CB8FFF8FFFFF6FE00127A0010000020A7
-:1010E00000093D001400002070B505460E4600249E
-:1010F00000BF0CE06000295A6548FFF7ABFD00BF58
-:1011000080216348FFF7B8FE0028F9D0641CB44280
-:10111000F0DB70BD10B55F4801885F48FFF7E4FF62
-:10112000642000F048F95D4801885D48FFF7DCFF66
-:10113000642000F040F910BD70B505460E4600244D
-:1011400000BF11E000BF20215148FFF795FE0028A5
-:10115000F9D020214E48FFF799FE4D48FFF77EFD5C
-:1011600061006852601CC4B2B442EBDB012070BD68
-:1011700010B506221521484800F09CF84549088022
-:1011800007221521464800F095F8444908800822B6
-:101190001521444800F08EF84349088009221521A2
-:1011A000424800F087F84249088010BD00B589B078
-:1011B00001214804FFF763FB01218803FFF76DFB62
-:1011C0000C200790022108A80170002181700121E4
-:1011D000C1700321417007A90920C006FFF71CF860
-:1011E000012202210920C006FFF79FF801220321F6
-:1011F0000920C006FFF799F84B20C001019000209C
-:1012000006900C200590002004900390029001A904
-:101210001F48FFF7CFFB01211D48FFF764FC402169
-:101220001B48FFF733FE09B000BD8CB0FFF7BEFFCF
-:10123000FFF79EFF09E018216846FFF77DFF68462B
-:101240008089012801D1FFF765FFF4E710B5114847
-:1012500001881148FFF748FF322000F0ACF80F4832
-:1012600001880F48FFF740FF322000F0A4F809483A
-:1012700001880948FFF738FF322000F09CF8074842
-:1012800001880748FFF730FF322000F094F810BDC6
-:1012900000380140240000202C00002026000020FF
-:1012A000400000205400002028000020660000209C
-:1012B0002A00002070B504460D4621460020552323
-:1012C0002380A41CAA232380A41C2646A41C102D22
-:1012D00008D0132D02D0152D0CD107E01323238045
-:1012E000A41C08E010232380A41C04E01523238001
-:1012F000A41C00E000BF00BFFF232380A41C130038
-:1013000000F04BF90A3B063B0C3B3B121D29323BDC
-:1013100001232380A41C431C98B22FE003232380C5
-:10132000A41C431C98B229E002232380A41C01239F
-:101330002380A41C2380A41CC31C98B21EE002239B
-:101340002380A41C01232380A41C00232380A41C2D
-:10135000C31C98B212E003232380A41C0123238022
-:10136000A41C831C98B209E003232380A41C00233F
-:101370002380A41C831C98B200E000BF00BF0023A0
-:101380002380A41CFF232380A41CC31C3380C31D03
-:1013900098B20C4670BD0000002106E00A2100BF93
-:1013A0000A464B1E99B2002AFAD10246431E98B251
-:1013B000002AF3D17047002106E0064900BF0A4623
-:1013C0004B1E99B2002AFAD10246431E98B2002A57
-:1013D000F3D17047E02E0000002203098B422CD38A
+:1010E00000093D00140000208CB000F09FF800F0D3
+:1010F00042F909E01821684600F0EDF86846808959
+:10110000012801D100F003F9F4E770B504460D465B
+:101110002146002055232380A41CAA232380A41C3D
+:101120002646A41C102D08D0132D02D0152D0CD14D
+:1011300007E013232380A41C08E010232380A41CB1
+:1011400004E015232380A41C00E000BF00BFFF23A0
+:101150002380A41C130000F020FA0A3B063B0C3B42
+:101160003B121D29323B01232380A41C431C98B24F
+:101170002FE003232380A41C431C98B229E0022300
+:101180002380A41C01232380A41C2380A41CC31C33
+:1011900098B21EE002232380A41C01232380A41CF8
+:1011A00000232380A41CC31C98B212E003232380D5
+:1011B000A41C01232380A41C831C98B209E00323F0
+:1011C0002380A41C00232380A41C831C98B200E06D
+:1011D00000BF00BF00232380A41CFF232380A41C86
+:1011E000C31C3380C31D98B20C4670BD002106E0BD
+:1011F0000A2100BF0A464B1E99B2002AFAD10246C4
+:10120000431E98B2002AF3D17047002106E0064938
+:1012100000BF0A464B1E99B2002AFAD10246431E6D
+:1012200098B2002AF3D17047E02E000000B589B0D3
+:1012300001214804FFF723FB01218803FFF72DFB61
+:101240000C200790022108A8017000218170012163
+:10125000C1700321417007A90920C006FEF7DCFF19
+:10126000012202210920C006FFF75FF801220321B5
+:101270000920C006FFF759F84B20C001019000205B
+:1012800006900C200590002004900390029001A984
+:101290004848FFF78FFB01214648FFF724FC402117
+:1012A0004448FFF7F3FD09B000BD70B505460E4692
+:1012B000002400BF0CE06000295A3E48FFF7CAFC3A
+:1012C00000BF80213B48FFF7D7FD0028F9D0641C00
+:1012D000B442F0DB70BD70B505460E46002400BF79
+:1012E00011E000BF20213348FFF7C6FD0028F9D0E8
+:1012F00020213048FFF7CAFD2E48FFF7AFFC610000
+:101300006852601CC4B2B442EBDB012070BD10B562
+:10131000294801882948FFF7C8FF6420FFF775FFB7
+:10132000274801882748FFF7C0FF6420FFF76DFFBB
+:1013300010BD10B5204801882048FFF7B6FF3220C5
+:10134000FFF763FF1E4801881E48FFF7AEFF3220FB
+:10135000FFF75BFF184801881848FFF7A6FF322007
+:10136000FFF753FF164801881648FFF79EFF32200B
+:10137000FFF74BFF10BD10B5062215210F48FFF7F0
+:10138000C4FE0D490880072215210E48FFF7BDFE57
+:101390000B490880082215210B48FFF7B6FE0B49C0
+:1013A0000880092215210A48FFF7AFFE0949088085
+:1013B00010BD000000380140240000202C00002057
+:1013C00026000020400000205400002028000020BB
+:1013D000660000202A000020002203098B422CD343
 :1013E000030A8B4211D300239C464EE003460B4375
 :1013F0003CD4002243088B4231D303098B421CD3D7
 :10140000030A8B4201D394463FE0C3098B4201D3C8
