+++ debug/nfc.d
@@ -1,0 +1,33 @@
+.\debug\nfc.o: user\nfc.c
+.\debug\nfc.o: .\inc\stm32f0xx.h
+.\debug\nfc.o: C:\Keil_v5\ARM\PACK\ARM\CMSIS\4.1.1\CMSIS\Include\core_cm0.h
+.\debug\nfc.o: C:\Keil_v5\ARM\ARMCC\bin\..\include\stdint.h
+.\debug\nfc.o: C:\Keil_v5\ARM\PACK\ARM\CMSIS\4.1.1\CMSIS\Include\core_cmInstr.h
+.\debug\nfc.o: C:\Keil_v5\ARM\PACK\ARM\CMSIS\4.1.1\CMSIS\Include\core_cmFunc.h
+.\debug\nfc.o: .\inc\system_stm32f0xx.h
+.\debug\nfc.o: .\Headers\stm32f0xx_conf.h
+.\debug\nfc.o: .\inc\stm32f0xx_cec.h
+.\debug\nfc.o: .\inc\stm32f0xx.h
+.\debug\nfc.o: .\inc\stm32f0xx_crc.h
+.\debug\nfc.o: .\inc\stm32f0xx_comp.h
+.\debug\nfc.o: .\inc\stm32f0xx_dac.h
+.\debug\nfc.o: .\inc\stm32f0xx_dbgmcu.h
+.\debug\nfc.o: .\inc\stm32f0xx_dma.h
+.\debug\nfc.o: .\inc\stm32f0xx_exti.h
+.\debug\nfc.o: .\inc\stm32f0xx_flash.h
+.\debug\nfc.o: .\inc\stm32f0xx_gpio.h
+.\debug\nfc.o: .\inc\stm32f0xx_syscfg.h
+.\debug\nfc.o: .\inc\stm32f0xx_i2c.h
+.\debug\nfc.o: .\inc\stm32f0xx_iwdg.h
+.\debug\nfc.o: .\inc\stm32f0xx_pwr.h
+.\debug\nfc.o: .\inc\stm32f0xx_rcc.h
+.\debug\nfc.o: .\inc\stm32f0xx_rtc.h
+.\debug\nfc.o: .\inc\stm32f0xx_spi.h
+.\debug\nfc.o: .\inc\stm32f0xx_tim.h
+.\debug\nfc.o: .\inc\stm32f0xx_usart.h
+.\debug\nfc.o: .\inc\stm32f0xx_wwdg.h
+.\debug\nfc.o: .\inc\stm32f0xx_misc.h
+.\debug\nfc.o: C:\Keil_v5\ARM\ARMCC\bin\..\include\stdio.h
+.\debug\nfc.o: user\nfc.h
+.\debug\nfc.o: user\cmd.h
+.\debug\nfc.o: user\delay.h
